Gloss2Text2Speech/requirements.txt
accelerate==1.0.1
aiohappyeyeballs==2.4.4
aiohttp==3.10.11
aiosignal==1.3.1
altair==5.4.1
annotated-types==0.7.0
anyio==4.5.2
async-timeout==5.0.1
attrs==24.3.0
blinker==1.8.2
cachetools==5.5.1
certifi==2024.12.14
charset-normalizer==3.4.1
click==8.1.8
contourpy==1.1.1
cycler==0.12.1
datasets==3.1.0
dill==0.3.8
distro==1.9.0
evaluate==0.4.3
exceptiongroup==1.2.2
fasttext==0.9.3
filelock==3.16.1
fonttools==4.55.5
frozenlist==1.5.0
fsspec==2024.9.0
gitdb==4.0.12
GitPython==3.1.44
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
huggingface-hub==0.27.1
idna==3.10
importlib_resources==6.4.5
Jinja2==3.1.5
jiter==0.8.2
joblib==1.4.2
jsonschema==4.23.0
jsonschema-specifications==2023.12.1
kiwisolver==1.4.7
markdown-it-py==3.0.0
MarkupSafe==2.1.5
matplotlib==3.7.5
mdurl==0.1.2
mpmath==1.3.0
multidict==6.1.0
multiprocess==0.70.16
narwhals==1.24.0
networkx==3.1
numpy==1.21.0
openai==1.60.2
packaging==24.2
pandas==2.0.3
peft==0.13.2
pickle-mixin==1.0.2
pillow==10.4.0
pkgutil_resolve_name==1.3.10
portalocker==3.0.0
propcache==0.2.0
protobuf==5.29.3
psutil==6.1.1
pyarrow==17.0.0
PyAudio==0.2.14
pybind11==2.13.6
pydantic==2.10.6
pydantic_core==2.27.2
pydeck==0.9.1
Pygments==2.19.1
pyparsing==3.1.4
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
PyYAML==6.0.2
referencing==0.35.1
regex==2024.11.6
requests==2.32.3
rich==13.9.4
rpds-py==0.20.1
safetensors==0.5.2
scikit-learn==1.3.2
scipy==1.10.1
six==1.17.0
smmap==5.0.2
sniffio==1.3.1
streamlit==1.40.1
sympy==1.13.3
tenacity==9.0.0
threadpoolctl==3.5.0
tokenizers==0.20.3
toml==0.10.2
torch==2.4.1
torch-lr-finder==0.2.2
tornado==6.4.2
tqdm==4.67.1
transformers==4.46.3
typing_extensions==4.12.2
tzdata==2025.1
urllib3==2.2.3
watchdog==4.0.2
xxhash==3.5.0
yarl==1.15.2
zipp==3.20.2
